Updates to get ffmpeg reading and writing working for the Matroska container. Fixed issues with many of the codecs by populating the pts member of the frame with a monotonically increasing frame number. The video duration is however still read incorrectly for video encoded with raw codec in at least the matroska container.

+++ include/flitr/ffmpeg_writer.h
@@ -40,10 +40,12 @@ extern "C" {
 # include <swscale.h>
 # include <libavutil/mathematics.h>
 # include <libavcodec/avcodec.h>	
+#include <libavutil/pixdesc.h>
 #else
 # include <avformat.h>
 # include <libavutil/mathematics.h>
 # include <libavcodec/avcodec.h>
+#include <libavutil/pixdesc.h>
 #endif
 }
 
@@ -125,7 +127,7 @@ private:
     std::string SaveFileName_;
     AVRational FrameRate_;
     /// Holds the number of frames written to disk.
-    uint32_t WrittenFrameCount_;
+    uint64_t WrittenFrameCount_;
 
 
     uint8_t* VideoEncodeBuffer_;
